Again

+++ torchvision/csrc/cpu/vision_cpu.h
@@ -89,12 +89,12 @@ at::Tensor DCN_forward_cpu(
     const at::Tensor& input,
     const at::Tensor& offset,
     const at::Tensor& weights,
-    const int64_t stride,
-    const int64_t padding,
-    const int64_t dilation,
-    const int64_t groups,
-    const int64_t deformable_groups,
-    const int64_t im2col_step);
+    const int stride,
+    const int padding,
+    const int dilation,
+    const int groups,
+    const int deformable_groups,
+    const int im2col_step);
 
 at::Tensor DCN_backward_cpu(
     const at::Tensor& grad, const at::Tensor& input);
